Fixes #71 Updated version reference of Microsoft.Extensions.Caching.Memory to 2.1.2 to match Microsoft.AspNetCore.App 2.1 version restrictions Added additional web sample project targeting .net core 2.1 (Long Term Support version) Removed unnecessary package references because of dependent packages Ensured that an exe referencing Microsoft.Extensions.Caching.Memory --version 2.2.0 will execute correctly Executed unit tests to ensure passing
