+++ CacheDatabaseQueriesApiSample.netCore2.1LTS/CacheDatabaseQueriesApiSample.netCore2.1LTS.csproj
@@ -1,0 +1,21 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.1.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\LazyCache.AspNetCore\LazyCache.AspNetCore.csproj" />
+    <ProjectReference Include="..\LazyCache\LazyCache.csproj" />
+  </ItemGroup>
+
+</Project>
