+++ CacheDatabaseQueriesApiSample.netCore2.2/Controllers/DbQueriesController.cs
@@ -1,0 +1,15 @@
+﻿using CacheDatabaseQueriesApiSample;
+using Microsoft.AspNetCore.Mvc;
+
+namespace ApiAsyncCachingSample.Controllers
+{
+    public class DbQueriesController : Controller
+    {
+        [HttpGet]
+        [Route("api/dbQueries")]
+        public int GetDatabaseRequestCounter()
+        {
+            return DbTimeContext.DatabaseRequestCounter();
+        }
+    }
+}
