--- CacheDatabaseQueriesApiSample/Controllers/DbQueriesController.cs
@@ -1,15 +1,0 @@
-﻿using CacheDatabaseQueriesApiSample;
-using Microsoft.AspNetCore.Mvc;
-
-namespace ApiAsyncCachingSample.Controllers
-{
-    public class DbQueriesController : Controller
-    {
-        [HttpGet]
-        [Route("api/dbQueries")]
-        public int GetDatabaseRequestCounter()
-        {
-            return DbTimeContext.DatabaseRequestCounter();
-        }
-    }
-}
