--- CacheDatabaseQueriesApiSample/Properties/launchSettings.json
@@ -1,29 +1,0 @@
-{
-  "iisSettings": {
-    "windowsAuthentication": false,
-    "anonymousAuthentication": true,
-    "iisExpress": {
-      "applicationUrl": "http://localhost:52671/",
-      "sslPort": 0
-    }
-  },
-  "profiles": {
-    "IIS Express": {
-        "commandName": "IISExpress",
-        "launchBrowser": true,
-        "launchUrl": "/",
-        "environmentVariables": {
-            "ASPNETCORE_ENVIRONMENT": "Development"
-        }
-    },
-    "CacheDatabaseQueriesApiSample": {
-      "commandName": "Project",
-      "launchBrowser": true,
-      "launchUrl": "/",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
-      },
-      "applicationUrl": "http://localhost:52672/"
-    }
-  }
-}
