+++ LazyCache.AspNetCore/LazyCache.AspNetCore.csproj
@@ -24,9 +24,7 @@
 
 
   <ItemGroup>
-    <PackageReference Include="microsoft.extensions.dependencyinjection.abstractions" Version="2.2.0" />
-    <PackageReference Include="microsoft.extensions.caching.abstractions" Version="2.2.0" />
-    <PackageReference Include="microsoft.extensions.caching.memory" Version="2.2.0" />
+    <PackageReference Include="microsoft.extensions.caching.memory" Version="2.1.2" />
     <ProjectReference Include="..\LazyCache\LazyCache.csproj" />
   </ItemGroup>
