+++ LazyCache.Ninject/LazyCache.Ninject.csproj
@@ -23,8 +23,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="microsoft.extensions.caching.abstractions" Version="2.2.0" />
-    <PackageReference Include="microsoft.extensions.caching.memory" Version="2.2.0" />
+    <PackageReference Include="microsoft.extensions.caching.memory" Version="2.1.2" />
     <PackageReference Include="Ninject" Version="3.3.4" />
     <ProjectReference Include="..\LazyCache\LazyCache.csproj" />
   </ItemGroup>
